Add a separate live set, to track ownership state within a block.
The GEN set was being overloaded, to track both the fact that ownership was generated (GEN)
within a block and the current ownership state (LIVE).  If ownership is both created and consumed
in a single block, then these two quantities differ: GEN should be true while LIVE should
be false.  This patch adds the LIVE set and tracks it separately.

I also reworked backward propagation to propagate deadness (0s in the OUT set, potentially
requiring the insertion of autoDestroy calls) backward through the flow graph.  This will
allow locally-allocated memory to be reclaimed as early as possible.  This is a
work-in-progress, because it is now too aggressive.  (Not even hello runs correctly.)

If an owned variable is created, it is an error to insert an autoDestroy call before its
last use.  So I have to add code to track uses and propagate these backward:
 -- An autoDestroy should be inserted for each symbol that is live at the end of a block
    and whose last use lies within that block.
 -- A consumer that is in the shadow of a (non-consuming) use must not consume the symbol
    (an autoCopy must be insterted).

Factored out the loops that set or reset bits for all members of an alias list.  Note that
this apparently doubles the loop overhead in some places.  But when optimization is turned
on, GCC is expected to inline and coalesce these loops.  So in the end, the code should be
more readable without giving up any performance.