Add and track last use to avoid premature destruction.
### b/compiler/include/bitVec.h
Added += and -= operators.

### b/compiler/passes/insertAutoCopyAutoDestroy.cpp
Renamed GEN as PROD and KILL as CONS.
Added the USE set.
Added the isUsed() predicate to test when a variable is read.  Aliases through references
are not tracked in this implementation.
Added processUser(), and call this from computeTransitions().
Added backwardFlowUse(), to determine the block (on each branch) in which symbols are
used.
Modified the in-block equation in forwardFlowOwnership() to use the USE set to consume
ownership.  Since an autoDestroy call will be added for each symbol that is owned and last
used in that block, a USE bit actually translates to consumption.
Commented out backwardFlowOwnership().  It is no longer needed.