Improved last-use computation.
Some tests were failing because we were failing to add autoDestroy calls at the end of the
function if the variable was only every set and never read (classes/diten/test_destructor
etc.)  I modified isUsed to also treat defs as uses, which simplifed it a lot.

In backwardFlowUse(), changed "out" from a pointer to a reference.  Modified the
postprocessing step to recompute IN as the union of the OUTs of preceding blocks.  This
sets USE to true in a block where IN is true and OUT is false (i.e. the last use on a
branch can actually be no use at all).

The added block at the end of forwardFlowOwnership may thus be moot.

27 => 18 regressions