Updates
Copy uses getGID and getUID (I finalized the naming).  Since allowing copy to
take a destination directory would be easier with the string-related operations
basename and joinPath, decided to hold off on that functionality and error out
if dest was a directory, so updated the comment accordingly and left a FUTURE
note, but put some structure in place for it.  Made copyFile check on the
existence of the source file (since it is going to open it and copy its
contents).  Made copyFile check that neither its src or dest were directories.
Made copyFile test a little more robust (but it still needs some work).