Ensure last-* exists when configuration is run for first time.
Previously, nightly tests that only used nightly_email.pl would report a number
of a failures but fail to include the actual failure. This was due to the
last-* file not existing, so the comm commands failed. This would not have
happened for tests that use nightly, since it was ensuring the last-* file
existed.

Update nightly_email.pl to create an empty last-* file if it does not already
exist. This way, any errors that show up on the first run will be reported as
new errors.