Update printchplenv 3rd party vars to work, even when CHPL_HOME is not set.
Previously, if CHPL_HOME was unset, printchplenv would fail with an OSError and
stack trace when trying to run third-party-pkgs. This is due to CHPL_HOME
defaulting to empty string if not set in environment, then calculating the path
to third-party-pkgs script using CHPL_HOME as part of the path (e.g. the path
calculation was: '' + '/util/chplenv/third-party-pkgs').

Update third-party-pkgs path calculation to use path relative to __file__ (aka
the running printchplenv script) when CHPL_HOME is unset.

To reproduce the error on master (before this change), use:

```bash
( unset CHPL_HOME ; ./util/printchplenv --make )
```