Add a check in the bulk optimization to avoid memcpy(a,a)
Continuing PR #1410, it seems that sometimes two Chapel
arrays can have the same underlying data even if the _value
is different. I think this is from array slicing or aliasing,
and the symptom is that the madness tests fail with the
previous patch.

In this patch, I adjust the module code for bulk optimization
to directly check for src==dst before doing the call to
comm get. That resolves the errors with the madness tests.