Repair problems in yesterday's third-party integration changes.
PR #1343 made big changes in how third-party packages are integrated
into Chapel, but unfortunately it also broke some things.  Here I'm
fixing those errors.

The major problem was that we ended up with a mismatch in the unique
configuration path for gmp and re2, between the chpl_<pkg>.py scripts
which needed that path to see if an install had already been done, and
(effectively) the chpl_3p_<pkg>_configs.py scripts which specified that
path.  There were actually two problems.  One was that we didn't have
the uniq-cfg-path set at all on a clean checkout, because printchplenv
was only running the chpl_3p_*_configs.py scripts for the third-party
packages that had already been built.  The other was that in some cases
we were using the LCD of the target architecture and in others we were
not.

The major change here is that now we set the uniq-cfg-path make vars
in printchplenv for all third-party packages, whether or not we've built
them yet.  (We still only set the link-args variables for those that
have been built, though.)

The other change is in third_party_utils.py, where we use the same logic
that printchplenv does to decide whether or not to ask for the LCD of
the target architecture when building the default uniq-cfg-path.

In chpl_gmp.py and chpl_re2.py, we now get the uniq-cfg-path from the
corresponding chpl_3p_<pkg>_configs.py scripts instead of building it
ourselves.  This ensures that the installs and the subsequent checks for
installs use the same uniq-cfg-path.

And finally, now chpl_3p_gmp_configs.py returns the right link-args when
CHPL_GMP=system, with a little bit of help from printchplenv.  It didn't
do this before.  That problem hadn't yet been noticed but I spotted it
when I was working on the other stuff here.