Cleanup how leader and standalone iters are marked for inlining
Adds helper routines that check if a function is a specific type of iterator
(leader, standalone, follower.) These functions are then used to find leader
and standalone iterators and mark them for inlining. This separates the logic
of checking if a fn is a particular iterator from whatever you want to do with
the iterator.

This also moves the standalone marking up with the leader marking.

This is a peeled off commit from my vectorize-follower-loops patch (which also
adds and uses an isFollowerIterator function())