Add version detection macros to runtime/include/chpl-comp-detect-macros.h
runtime/include/chpl-comp-detect-macros.h contains code to detect which backend
compiler is being used. This adds version detection macros for all our backend
compilers primarily for feature detection. These are required for the upcoming
vectorize-loop-followers patch that attaches the ivdep pragma to loops in
follower iterators. It is being added separately so the links/info at the end
of this commit message can be easily found in the commit history.

This does not provide a single macro for all compilers but instead provides
unique macros for each compiler (since there's no uniform versioning scheme
between compilers and you want to be explicit about which compiler you're
checking.) This primarily serves as a means to avoid having to look up how each
compiler stores it's version number since the information isn't that easy to
find.

Version number schemes:
- Cray  : _RELEASE_MAJOR, _RELEASE_MINOR
- Intel : __INTEL_COMPILER
- Clang : __clang_major__, __clang_minor__, __clang_patchlevel__
- GCC   : __GNUC__, __GNUC_MINOR__, __GNUC_PATCHLEVEL__
- PGI   : __PGIC__, __PGIC_MINOR__, __PGIC_PATCHLEVEL__

These are the links I used for each compiler:
- Cray  : http://docs.cray.com/books/S-2179-83/S-2179-83.pdf
- Intel : https://software.intel.com/sites/products/documentation/doclib/iss/2013/compiler/cpp-lin/GUID-603F2191-6FC2-4EE6-8B94-55C7CAD72D2E.htm
- Clang : http://clang.llvm.org/docs/LanguageExtensions.html#builtin-macros
- GCC   : https://gcc.gnu.org/onlinedocs/cpp/Common-Predefined-Macros.html
- PGI   : http://sourceforge.net/p/predef/wiki/Compilers/
- Misc  : http://nadeausoftware.com/articles/2012/10/c_c_tip_how_detect_compiler_name_and_version_using_compiler_predefined_macros

Confirmed by checking the predefined macros for each compiler:
- Cray  : cc -Wp,-list_final_macros -E <test c file (can't be /dev/null)>
- Intel : icc -dM -E -x c /dev/null
- Clang : clang -dM -E -x c /dev/null
- GCC   : gcc -dM -E -x c /dev/null
- PGI   : pgcc  -dM -E

Note that the major version for cray is _RELEASE_MAJOR *NOT* _RELEASE. _RELEASE
used to be what the docs specified, but there was a bug where _RELEASE reported
the wrong version for the c compiler. Since 8.3 the docs (and a closed bug) say
to use _RELEASE_MAJOR. The _RELEASE_MAJOR macro has been around for years so we
can rely on it, it just wasn't documented.

Note that while the clang version macros are provided, the clang website
insists that they are *NOT* to be used for feature detection as "different
vendors use different numbering schemes."