Update AstPrintDocs visitor to correctly handle "no doc" pragmas on submodules.
Previously, the "no doc" pragma was ignored when traversing submodules. The
module documentation would not be produced (correcty), but traversal would not
stop so anything inside the submodule would be documented (incorrect).

Update the visitor to check if the module has the "no doc" pragma, and, if so,
stop the AST traversal and avoid documenting the submodule.