Bug fix: Stash a pristine copy of standalone iters like we do for leaders
During review Vass noticed that we should also be stashing a pristine copy of
the standalone iters.

A better description from Vass:
"
With the introduction of standalone iterators, the forall intent
machinery has been extended to apply to them, too. Standalone iterators
are treated much like leader iterators for forall intents. (The function
names and comments in implementForallIntents.cpp are still heavy in
leader-follower terminology - for historical reasons.)

This change fixes an omission in the above - we need to "stash pristine
copies" of standalone iterators just like we do leader iterators. The
omission has not caused us trouble yet perhaps because we have not used
standalone iterators in some ways that we have leaders.
"