Disallow arguments of type string for extern procs
This change makes it a compiler error to use a string in the
prototype of a extern proc, and recommends the use of c_string instead.

Using string can cause problems due to wideness when used with extern
procedures. Additionally, when 'strings as records' goes in, users will
probably never want the actual string record to be passed out (and they
will get a C compile time error if they expected a c_string).

If it becomes necessary we can add a pragma that disables this check on
a function.