Move ScopeResolve.cpp:findOuterLoop() to LoopStmt::findEnclosingLoop()
findOuterLoop() was a static function that lived inside of scope resolve. I need
this function outside of scope resolve for my upcoming vectorize-loop-followers
patch. I also think it will be ideal to use in Loop::insertBefore() in loop
invariant code motion, though I haven't tested it there yet.

This just moves it's definition as a static function in scope resolve to a
static method on LoopStmt. It also renames it from findOuterLoop to
findEnclosingLoop because findOuterLoop implied (at least to me) that it would
find the outer most loop in a nest of loops, which is not accurate.