Added code to insert autoDestroy calls for owned symbols at scope exit.
Added and populated the EXIT bit set.  A bit is set in EXIT[i] for each symbol that goes
out of scope in block i.  Ownership analysis treates EXIT[i] being true as an indication
that ownership cannot propagate out of block i.

Since EXIT[i] being true forces OUT[i] to false for the corresponding symbol, the standard
equation for computing the to_cons set will show when ownership is forced to zero by EXIT,
and will insert an autoDestroy for the corresponding symbol.

The USE set is no longer used, but its computation is left in place, since it may be
useful in implementing xlvalue binding.