Add "unowned return value" optimization for getters.
[The idea and implementation of this commit are courtesy of Vass.]

Since getters effectively return a reference to an object that is owned elsewhere (that
is, the record or class containing the selected field), their return value can be treated
as unowned.  That makes it unnecessary for the getter itself to insert an autoCopy, and
pushes responsibility for that decision into the caller.

The caller only needs to insert an autocopy of the value returned from a getter if that
value is consumed (stored in a global, stored in a field or returned).  If the value is
only read and not consumed, then an autoCopy need not be inserted.  This results in a net
reduction in the number of autoCopies in those cases.

The implementation is straightforward:
  - Getter functions returning values of record type are marked with a new
    FLAG_RETURN_VALUE_IS_NOT_OWNED.
  - The return primitive in marked functions is not treated as a consumer.
  - The value returned by marked functions is treated as unowned.
The rest of the AMM implementation remains unchanged.  The normal AMM machinery takes care
of inserting autoCopies where they are needed.

This change altered (beneficially) the output of classes/diten/record_initialize, reducing
the number of calls to BBB.initialize() by 5.