Add isCoforallLoop() method
Currently coforall loops are just represented as ForLoops within the compiler.
This is a start at separating them. This is a peeled off commit from my
vectorize-loop-followers branch where I want a clean way to distinguish which
loops are really coforall loops.

Eventually CoforallLoop should be it's own class that shares a parent with
ForLoop. That's more work than I want to tackle this close to the release.
Having a new class would continue the work Mike started with the other loop
types and should allow us to retire PRIM_BLOCK_COFORALL and
FLAG_COFORALL_INDEX_VAR.