Replace fftw_complex a Chapel's complex(128)
This turned out to be trivial.  Nikhil ran into issues with it
originally, but that may have been back when he was still using
explicit c_ptr()-style argument passing (which is probably more
sensitive to type mismatches).

I replaced fftw_complex with complex(128) in FFTW.chpl's interfaces
and testFFTW.chpl, got rid of the utility routines, and rewrote
the testFFTW.chpl code itself to use the native types and operations.