Improve alias tracking for AMM.
After revising AMM so that it was inserting autoDestroys at the end of the scope in which
a symbol is declared, there were some early deletion (read-after-free) errors the cropped
up in about a dozen tests:
  distributions/robust/arithmetic/basics/test_array_of_distributed_array2.chpl
  release/examples/benchmarks/shootout/meteor-fast.chpl
  spec/marybeth/sort.chpl
  studies/cholesky/jglewis/version2/dataflow/test_dataflow_cholesky.chpl
  studies/cholesky/jglewis/test_dataflow_cholesky.chpl
  studies/hpcc/RA/bradc/ra-bradc-old1.chpl
  studies/hpcc/RA/marybeth/ra3.chpl
  studies/shootout/fannkuch-redux/sungeun/fannkuch-redux.chpl
  studies/shootout/k-nucleotide/bharshbarg/knucleotide-chaining.chpl
  studies/shootout/k-nucleotide/bharshbarg/knucleotide-associative.chpl
  studies/shootout/meteor/kbrady/meteor-parallel-alt.chpl

Examination of the generated code revealed that ownership was being shared between symbols
declared in different scopes and that exiting the inner scope was causing one member of
this alias set to be destroyed.  Later references to other members of the alias set are
then (obviously) invalid.

The cure is to limit the membership of an alias clique to symbols declared in the same
scope.  Ownership can be transferred from a clique in an inner scope to an outer scope,
but not vice versa.

##DETAILS
*Added the utility functions getScopeBlock() and isParentExpr().
*Revised createAlias() to ensure that when alias pairs are created, both symbols are declared in the same scope.
*Used getScopeBlock instead of inline code, where appropriate.
*Revised processBitwiseCopy() to only transfer ownership to a symbol in an outer scope now.

[x] Tested on the 12 failing test cases.