Make copy ignore ENOENT on destination in copy before copyFile is executed
It doesn't matter if the destination didn't exist before, we're going to create
it when we copy to that location.  Other errors would matter, though.

Also, updates the comment in the copyFile test, stating my assumption that umask
will not be set to only give user read permissions and nothing else.