Add three futures on a segfault that occurs under very specific circumstances.
This bug will occur on functions that fulfill all the following conditions:
- Utilizes a file and channel for reading and/or writing
- Sometimes exits before the utilization of said file and channel (will only
  segfault in the early exiting case)
- When it exits, returns a value via an out argument instead of a return
  statement.

If reading is done, the program will only segfault with --no-devel thrown or
CHPL_DEVELOPER set to false.  If writing is done, the program will only
segfault with --devel thrown or CHPL_DEVELOPER set to true.  When both are
present, the program will segfault under

Tom has tested that this bug is not present in his AMM branch, so this bug
should be resolved when that is committed.