Isolate, deprecate and remove old-style AST collector functions, pass 1.
Created a new header file oldCollectors.h, and moved old-style collector functions there.
This is a first step in deprecating these in favor of the STL-based collector functions.

After making this change, it was necessary to either convert all the collector calls in a
source file to use the STL-based collectors or to add an #include "oldCollectors.h"
directive.

I used an ad hoc selection criterion which way to go (upgrade or not) for each source
file, which was roughly equivalent to "don't upgrade if there are 3 or more compile-time
errors in that file".

In some places it was necessary to rename the iterator variable, because the STL version
of for_vector publishes a declaration of iterator variable in the enclosing scope.

[x] Tested on examples (std)