Removed parSafe param arguments from each call
In documenting this interface, it seemed like overkill to permit the
user to be able to override the parallel safety of the class on each
method and in a survey of users, nobody objected to retiring this
part of the interface (and a few developers rejoiced at the ensuing
simplicity).

The main downside of doing this is that we have some methods call
others in their implementation, and to avoid grabbing a lock twice,
they used these optional arguments to avoid the locks.  But that
doesn't seem like a good reason to make those part of the user's
interface, so here I merely refactor the guts that are protected
by locks into "private" methods that don't take the locks.

In discussing this feature, it was pointed out that the param-ness
of parSafe also feels a little ridiculously heavyweight/static
given that it's just a conditional around a few calls.  So if
a user did want to change the parallel safety of their random
stream over its lifetime, my suggestion would be to change that
from a param to a var/const so that it can be changed during the
object's lifetime rather than on a call-by-call basis.