Since I check that copyFile doesn't copy permissions, figured I should also check that copyMode doesn't copy contents, so I modified the test to output the contents of both files.
Realized I wasn't taking the metadata argument into account, so I surrounded
the uid and gid portions with it.  Will add access time and time of last
modification later.  Then altered the copy test to send in metadata=true
(instead of altering it to skip checking the uid and gid)

Also realized that the comment on the dest argument for both versions of copy
did not reflect the current "don't allow directories" state of things.  So I
updated that.