Moved FFTW module to modules/standard
[trivial, not reviewed; Nikhil OK'd]

Nikhil and I think that the current draft of the FFTW module is ready
for prime-time and are moving it to modules/standard.  I'll ask for a
second review from someone else on the core team before the release
goes out, but want to get the documentation a bit more up-to-snuff
before doing so (since it'll help, and can be reviewed at the same
time).