Proposed split of FFTW into FFTW and FFTW_MT
Spltting the FFTW module into a single- and multi-threaded
component is nice in that it permits the multi-threaded
version to initialize itself at module initialization time.
It also permits each file to list its own "requires"
dependence if I can get that agreed upon and committed in
time to make use of it.  Committing this for Nikhil's
purusal.