Add CHPL_PRAGMA_IVDEP macro for backend compilers
In order to help the backend compiler vectorize we're going to be attaching the
ivdep (ignore vector dependency) pragma to 'follower' loops. The various
backend compilers have different versions of the pragma and we have to be
careful about which version of the compiler started supporting the pragma.

The chapel compiler will always generate the CHPL_PRAGMA_IVDEP, but the actual
definition of that macro depends on the backend compiler and backend compiler
version.

The specific pragmas were found by looking at the reference manuals for each
compiler. The version numbers were picked by looking at archived reference
manuals for many of the compilers. GCC only recently started supporting and
ivdep pragma with version 4.9 and clang does not appear to have one at least as
of version 3.6

Cray:
 - The cray compiler has supported the ivdep pragma for a long time. We could
   get away without a version check since CCE isn't released outside of the
   PrgEnv but it doesn't hurt.

Intel:
 - Intel has also supported the ivdep pragma for a long time. The earliest
   reference manual I found was for version 9. It's possible earlier compilers
   supported the ivdep pragma too but version 9 was released in 2005 and that's
   more than enough prior releases for us.

Clang:
 - Surprisingly clang doesn't support the ivdep pragma. They have a vectorize
   pragma, but that just manually enables/disables vectorization on a specific
   loop, it doesn't provide the compiler with any more information so it
   doesn't give us anything. I did see some RFCs for an ivdep pragma on Clang's
   mailing lists and we know the llvm has hooks for accepting this kind of
   vectorization hint so it will only be a matter of time until there is a way
   to tell Clang's C compiler.

PGI:
 - Pgi has an ivdep pragma, but only for fortran. They do have a nodepchk
   pragma, which as stated in their reference manual has identical semantics to
   ivdep, but it is supported for C/C++ as well. PGI has supported this since
   at least 2006.

GCC:
 - GCC only recently added supported for the ivdep pragma. It is "GCC ivdep"
   instead of just "ivdep".