Update chpldoc to work when CHPL_HOME is moved or changed.
Previously, the virtualenv install for the python packages hard coded the paths
at the time of install. If this was built into the module rpm, for example, the
path to CHPL_HOME would change potentially with each install of the module and
chpldoc would not work.

Update the python package install to use virtualenv --relocatable [1], which
updates all the hard coded paths to be relative. One of the caveats when using
the --relocatable flag, is the activate scripts no longer work. Fortunately,
the only required things the activate scripts do is set VIRTUAL_ENV in the
environment and update PATH to have the python "binaries" from the virtualenv
at the beginning. So, also update the docs() pass to set VIRTUAL_ENV and PATH
environment variables instead of calling `. path/to/activate`.

I verified this by copying my chapel working directory with a build of chpldoc
somewhere else on my file system, deleting the install dir in
third-party/chpldoc-venv/ from the original workspace, and then running chpldoc
in the new directory. It worked as expected for `make docs`.