Attach pragma ivdep to loops inside of 'follower' iterators
Attach pragma ivdep to loops we know are order independent so that our
generated code will vectorize better. ivdep asserts to the c compiler that it
can ignore assumed vector dependencies between loop iterations. This can help
the backend compiler with vectorization. It will either allow the compiler to
vectorize a loop it wasn't sure was safe previously, or it will allow the
compiler to vectorize without having to clone the loop. Previously, if the
backend compiler wasn't sure if there were dependencies it might try to
vectorize but it would generate a vectorized and non-vectorized version of the
loop and decide at execution if the vectorized loop was safe.

During function resolution we find all yielding loops inside of follower and
standalone iterators and mark them as order independent. The tricky part is
making sure we maintain order independence correctly during iterator lowering.
If a loop is order independent we codegen CHPL_PRAGMA_IVDEP which is defined
for each target compiler. Most target compilers support a "pragma ivdep". Clang
is the only target compiling lacking this pragma, and GCC only recently added
it with their 4.9 release.

Some code details:
 - Add code in LoopStmt to get/set if the loop is order independent

 - Add code generation for loops that are order independent (we generate
   CHPL_PRAGMA_IVDEP which is defined in the runtime for each target compiler)

 - Mark loops inside of follower and standalone iterators during resolution

 - Unmark loops that can't be inlined (build an advance function) since the
   advance function is 're-entrant' and must be called in order

 - Mark loops in the iterator body when iterators are being directly inlined

 - Mark loops correctly for zippered iterators
   - A zippered iterator will only be order independent if all of it's iters
     will be inlined (can't have any advance functions) and it was marked order
     independent before zippering or all of it's iters were order independent
   - This required exposing the function that checks if an iter that is
     zippered can be inlined (isSingleLoopIterator) and refactoring some
     reporting code so it doesn't fire every time isSingleLoopIterator() is
     called

 - Add flag to control emitting CHPL_PRAGMA_IVDEP. It is just --vectorize
   instead of --mark-order-independent-loops or something since this flag will
   control more in the future and better expresses the end goal of the flag. A
   flag to report which loops are order independent was also added. The test
   that checks --help was updated as well as the man page.

 - Note that the analysis to determine which loops are order independent is
   always run, it's only at codegen time that we check if we should emit the
   pragma. This is only because I think the analysis might be used for other
   things in the future and it adds checks in ~6 places that just make the code
   harder to read. It should be really cheap to do this analysis, but if we see
   a compiler performance hit, it's easy enough to only do the analysis when
   the flag is set.