Make these futures only run under valgrind.
While the segfault may or may not occur depending on whether --devel is
included, the incorrect free will always be visible to valgrind if it is
present, so remove the .compopts.  Also, remove the .bads because I don't want
to write valgrind specific .bads.  Updated the futures to indicate that the
skipifs (to ensure it only runs with valgrind) should be removed once the tests
pass there.