Add tests for attaching pragma ivdep to 'follower' loops
This adds a pretty extensive test suite. There are a few more tests I want to
add (an example with forall intents, a real world stream-like test, and a few
others) but this is pretty comprehensive.

This suite has several tests that define iterators to be invoked and checks
that CHPL_PRAGMA_IVDEP was emitted in the right places (and not emitted
anywhere else.) My initial approach grepped the generated code for the pragma
and the loops around it (and checked for the absence in other places) but that
quickly became unmaintainable.

For most tests (unless they have compopts that specify a main module) the
InvokeIterators module is used. All it does it setup some global vars and
invoke the leader/follower, standalone, and serial iterators. This avoid having
to repeat this code for all the tests.

I'd like to add a simple README that mentions what each test is testing since
it's not that easy to keep track of what cases have been covered.