Re-home paths taken from libtool .la files.
The paths in libtool .la files reflect where things were when the .la
files were built.  If the Chapel directory hierarchy is subsequently
moved (as for example when we install a module we've built), those paths
are no longer valid.  To Deal with this, when interpreting such paths
replace the part prior to the rightmost '/third-party/' component with
the current value of $CHPL_HOME.