Take Michael's comments into account.
Major change: made copyFile copy the contents in a binary, non-locking fashion,
and buffer the copying to 1024 bytes at a time.  The length of the buffer may
change over time (and Michael may correct that portion of the code, too).

Update the description of what "metadata" entails to be more specific (yes uid,
gid, last access, last modification, permissions independent of this arg).
Update description of behavior in case of an error to "will halt with an error"
from "will generate an error".