Repair retrieval of the hwloc install dir.
A side effect of the recent third-party integration changes is that a
given third-party package install directory will appear in the output
from 'compilelink --libraries' more than once if it is referred to both
explicitly by Chapel's user program Make infrastructure and implicitly
by at least one of the other third-party packages.  Hwloc is an example
of this.  Once things settle down I intend to clean this up, but in the
meantime it's breaking how maxTaskPar.prediff finds the lstopo utility
in order to collect hardware info.  Here we simplify this by getting the
hwloc install directory in a more straightforward way.