Update docs() pass to create output directory with 0777 permissions.
Previously, the permissions on the docs output directory were always 0700. This
makes the directory unreadable by other group members and users, by
default.

Update the permissions to be completely open by default. This is safe, because
the umask will still be applied. For example, when the umask is 0022, the
output directory will end up with 0755 permissions. If the umask is 0077, the
output directory will end up with 0700 permissions.