Simplify search for the final statement in a basic block.
Searching for the last non-flow statement cause autodestroys to be inserted ahead of
autoCopies inserted in the same block by this pass.  So we were killing a variable and
then copying it (classic read-after-free error).

This is half of the fix.  In blocks that do not end with a flow statement, it is still
possible for the autoDestroys to go in the wrong place.  We need to re-run BB analysis or
add the inserted autoCopies to the end of the basic block, to ensure that inserted
autodestroys follow them.

Now at 310 failures.