Add more checking for zero-length array copies
Check it in both array assignment and in doiBulkTransfer.
Zero-length arrays were having a different _value
but the same ddata, leading to the src!=dst assert
in chpl_memcpy firing without the earlier check in
doiBulkTransfer.

This patch adds the check both places in order to be conservative
and so that other distributions besides DefaultRectangular
don't have to worry about it.