Fix the bugs discovered in copy and friends.  Sigh.
My comparison of access and modification time failed on our linux32 testing
system because my error tolerance was not high enough.  The decimal place
represents fractions of a second, but we really only care about seconds (if
that).

Also, if the line read was shorter than the expected buffer, copyFile would
stop copying the file as a whole.  Whoops.  Just rely on EOF.