Recompute basic blocks before insertAutoDestroy().
Since insertAutoCopy may add statements to a block, BasicBlock::exprs can get
out-of-date.  In particular, since we look for the last statement in a basic block to
decide where to put autodestroy calls, we can get the wrong answer if we fail to keep the
exprs lists update or re-run basic block analysis.

The BasicBlock class would need to be reworked to support incremental updates to the
contents of a basic block.  Until that is done, it is easier to just re-compute basic
blocks before the insertAutoDestroy pass.

There is an implicit assumption that the flow graph is not changed by the insertion of
autoCopy calls, which I think could only happen if a formerly empty block had an autoCopy
call inserted into it.  On the other hand, autoCopy calls are inserted in front of a
statement containing a use of an unowned symbol, so that means it must be put into a block
that already contains something.  I think that a change in the flow graph is impossible.

Back down to 302 regressions.