Implement realPath for strings.
Basically just maps down to the C function realpath.  Add a test which ensures
that correctly modifying the here.cwd() path so that it contains some relative
paths in it but still refers to the same location and then sending that string
to realpath results in a string that matchs the here.cwd().