Refactoring module initialization
Based on feedback from Nikhil and Thomas, adding a config param
that permits a user to avoid FFTW_MT auto-initialization and
put the code that does the initialization into a routine so
that it can be invoked manually (this also has the nice effect
of cleaning up module initialization.

Noted the TODO that Nikhil and I discussed, but decided to wait
on based on user feedback, to make these routines methods on
locales if the "on all locales" nature is offensive to people.

Made the plan_with_nthreads() routine accept a sentinel value
by default that will cause it to use here.maxTaskPar on each
locale.  This is better than the old approach that assumed
that the value for locale #0 made sense for other locales as
well.