Put deref temp in the same scope as the original LHS.
Use se->var->defPoint as the insertion point instead of just before the
statement where the deref temp is defined.

Do this in both normalize.cpp and cullOverReferences.cpp.  I tried the change in
normalize first, but that had no effect on the test case I was looking at.  So I
implemented it in cullOverReferences as well.  I decided to leave in the change
in normalize, because it probably makes the code there more robust.

Down to 273 regressions.